Refactor `cli.ex` to extract module for setting config
We had a lot of private methods that all were essentially part of the same
operation - setting the `Config` struct - part of a module that had a lot going
on, so I extracted them all to a new module to group that behavior together and
hopefully make it a little easier to understand both modules now.
